refactor(settings): simplify per code-simplifier pass
Behaviour unchanged; tightens the new module's surface where
duplication was actually present, leaves alone places where CLAUDE.md
§4 ("≥3 times before extracting") didn't justify a helper.

Rust
- `OrgDetailsView::new(details, role)` + `IssuedInviteView::from`
  collapse the duplicated 8-field / 5-field literals in
  `read_org` / `update_org` / `invite_members` / `resend_invite`.
- `MemberStatus::as_str()` lives on the enum so the wire-label
  vocabulary stays next to the type; removes the file-local
  `member_status_label` helper in `pg_store`.
- `MAX_MEMBERS_PER_PAGE` retyped from `usize` to `u32` (its only
  callers compare/clamp against `u32`); deletes two `expect`-named-
  invariant blocks and lets `filter.per_page.clamp(1, MAX)` replace
  the `try_from + u32::min + .max(1)` chain.
- Removed the never-constructed `HttpError::InviteExpired` variant;
  folded its 410 arm into `Self::Org(OrgError::InviteExpired)`.

Frontend
- `SettingsPageHeader` extracted to the template module — the
  identical title + subtitle + right-slot block on both Settings
  pages is now defined once.
- `SettingsMembers` `RoleBadge` / `StatusPill` swap nested ternaries
  for `ROLE_STYLE` / `STATUS_STYLE` lookup tables.
- Filter tabs render from a `(["all","active","invited","expired"]
  as const).map(...)` loop instead of four near-identical calls.

Left untouched: `list_members` SQL + row-mapping loop (extracting
would either bloat the trait or force three round trips); core/
primitives (already minimal); `useOrg` hooks (already 3-5 lines
each); mock-backend (mostly contract code); test files.

+++ src/orgs/limits.rs
@@ -10,4 +10,9 @@ pub const INVITE_TTL: Duration = Duration::from_secs(7 * 24 * 60 * 60);
 pub const MAX_INVITE_BATCH: usize = 25;
 
 /// Hard ceiling on `?per_page=` for the members listing.
-pub const MAX_MEMBERS_PER_PAGE: usize = 50;
+///
+/// Typed as `u32` to match the wire-side `?per_page=` query
+/// parameter — the only callers compare or clamp against another
+/// `u32`, so a `usize` would just sprout `try_from` boilerplate at
+/// every seam.
+pub const MAX_MEMBERS_PER_PAGE: u32 = 50;
